requirements.txt
streamlit
pandas
difflib
time
sklearn.feature_extraction.text
sklearn.metrics.pairwise

